updated the code,extended some functionality
This code is a Bash script that automates the installation of important applications on a fresh Ubuntu installation. It includes a function to install snap packages for applications such as Discord, Notion, and Spotify, as well as basic tools like pip and npm. The script also updates the system and cleans up unnecessary packages.